Add directories in src
