+++ src/cli_terminal/mod.rs
@@ -1,0 +1,2 @@
+pub mod dummy_terminal;
+pub mod popup_terminal;
