+++ src/emulator/device/uart.rs
