+++ src/emulator/mod.rs
@@ -1,0 +1,4 @@
+pub mod cpu;
+pub mod memory;
+pub mod mmu;
+pub mod device;
