+++ src/wasm_terminal/mod.rs
@@ -1,0 +1,1 @@
+pub mod wasm_terminal;
